传

+++ src/main/java/com/quo/mapper/ProductMapper.java
@@ -50,7 +50,7 @@ public interface ProductMapper {
 	int saves(List<Product2> productList);
 	
 	//产品导出
-	List<ProductExport> getByPids(int[] pids);
+	List<ProductExport> getByPids(Long[] pids);
 	
 	@Select("select * from product where pid=#{pid}")
 	Product getProductByPid(Long pid);
